extract Category component
