+++ frontend/templates/contraceptives/tabs/Overview.tsx
@@ -3,6 +3,7 @@ import { Column, Row } from './StyledComponents';
 import { device } from '../../mediaSizes';
 import styled from 'styled-components';
 import PillRow from '../../../components/PillRow';
+import Category from '../../../components/Category';
 
 // Styling
 const CategoryRow = styled(Row)`
@@ -14,25 +15,6 @@ const CategoryRow = styled(Row)`
   width: 100%;
 `;
 
-const CategoryTitle = styled.h3`
-  color: #bebebe;
-  font-size: 1rem;
-  font-weight: normal;
-  margin: 0;
-`;
-
-const CategoryValue = styled.p`
-  color: #4b4b4b;
-  font-size: 1.5rem;
-  font-weight: 500;
-  margin: 0;
-  white-space: nowrap;
-
-  @media ${device.laptop} {
-    margin: 0.5rem 0;
-  }
-`;
-
 const Description = styled.p`
   color: #7c7c7c;
 
@@ -41,22 +23,6 @@ const Description = styled.p`
     width: 33vw;
   }
 `;
-
-// Components
-const Category = ({
-  title,
-  value,
-}: {
-  title: string;
-  value: string;
-}): ReactElement => {
-  return (
-    <Column>
-      <CategoryValue>{value}</CategoryValue>
-      <CategoryTitle>{title}</CategoryTitle>
-    </Column>
-  );
-};
 
 export interface OverviewProps {
   description: string;
